@@ -1,7 +1,7 @@
 // Systemd interface for non-Unix variants only
 
-//go:build windows || nacl || plan9
-// +build windows nacl plan9
+//go:build !unix
+// +build !unix
 
 package log
 
@@ -10,8 +10,12 @@ import (
 	"runtime"
 )
 
-// Enables systemd logs if configured or if auto detected
+// Enables systemd logs if configured or if auto-detected
 func startSystemdLog() bool {
 	log.Fatalf("--log-systemd not supported on %s platform", runtime.GOOS)
+	return false
+}
+
+func isJournalStream() bool {
 	return false
 }
