@@ -1,15 +1,16 @@
-//go:build !windows && !plan9 && !js
-// +build !windows,!plan9,!js
+//go:build unix
 
 // Package daemonize provides daemonization interface for Unix platforms.
 package daemonize
 
 import (
+	"fmt"
 	"os"
 	"strings"
 	"syscall"
 
 	"github.com/rclone/rclone/fs"
+	"golang.org/x/sys/unix"
 )
 
 // StartDaemon runs background twin of current process.
@@ -107,4 +108,21 @@ func argsToEnv(origArgs, origEnv []string) (args, env []string) {
 		}
 	}
 	return
+}
+
+// Check returns non nil if the daemon process has died
+func Check(daemon *os.Process) error {
+	var status unix.WaitStatus
+	wpid, err := unix.Wait4(daemon.Pid, &status, unix.WNOHANG, nil)
+	// fs.Debugf(nil, "wait4 returned wpid=%d, err=%v, status=%d", wpid, err, status)
+	if err != nil {
+		return err
+	}
+	if wpid == 0 {
+		return nil
+	}
+	if status.Exited() {
+		return fmt.Errorf("daemon exited with error code %d", status.ExitStatus())
+	}
+	return nil
 }
