@@ -116,13 +116,14 @@ func mountOptions(VFS *vfs.VFS, device string, mountpoint string, opt *mountlib.
 	for _, option := range opt.ExtraFlags {
 		options = append(options, option)
 	}
-	if runtime.GOOS == "darwin" {
+	// disabled in favor of ci.NoUnicodeNormalization -- see https://github.com/rclone/rclone/issues/7072#issuecomment-1604745556
+	/* if runtime.GOOS == "darwin" {
 		if !findOption("modules=iconv", options) {
 			iconv := "modules=iconv,from_code=UTF-8,to_code=UTF-8-MAC"
 			options = append(options, "-o", iconv)
 			fs.Debugf(nil, "Adding \"-o %s\" for macOS", iconv)
 		}
-	}
+	} */
 	return options
 }
 
