@@ -7,13 +7,8 @@ import (
 	"fmt"
 	"path/filepath"
 	"strings"
-	"time"
 
 	"github.com/moby/sys/mountinfo"
-)
-
-const (
-	pollInterval = 100 * time.Millisecond
 )
 
 // CheckMountEmpty checks if folder is not already a mountpoint.
@@ -80,19 +75,5 @@ func CheckMountReady(mountpoint string) error {
 	return fmt.Errorf(msg, mountpointAbs)
 }
 
-// WaitMountReady waits until mountpoint is mounted by rclone.
-func WaitMountReady(mountpoint string, timeout time.Duration) (err error) {
-	endTime := time.Now().Add(timeout)
-	for {
-		err = CheckMountReady(mountpoint)
-		delay := time.Until(endTime)
-		if err == nil || delay <= 0 {
-			break
-		}
-		if delay > pollInterval {
-			delay = pollInterval
-		}
-		time.Sleep(delay)
-	}
-	return
-}
+// CanCheckMountReady is set if CheckMountReady is functional
+var CanCheckMountReady = true
