@@ -3,10 +3,6 @@
 
 package mountlib
 
-import (
-	"time"
-)
-
 // CheckMountEmpty checks if mountpoint folder is empty.
 // On non-Linux unixes we list directory to ensure that.
 func CheckMountEmpty(mountpoint string) error {
@@ -19,9 +15,5 @@ func CheckMountReady(mountpoint string) error {
 	return nil
 }
 
-// WaitMountReady should wait until mountpoint is mounted by rclone.
-// The check is implemented only for Linux so we just sleep a little.
-func WaitMountReady(mountpoint string, timeout time.Duration) error {
-	time.Sleep(timeout)
-	return nil
-}
+// CanCheckMountReady is set if CheckMountReady is functional
+var CanCheckMountReady = false
