@@ -49,17 +49,18 @@ type Account struct {
 	// in http transport calls Read() after Do() returns on
 	// CancelRequest so this race can happen when it apparently
 	// shouldn't.
-	mu      sync.Mutex // mutex protects these values
-	in      io.Reader
-	ctx     context.Context // current context for transfer - may change
-	ci      *fs.ConfigInfo
-	origIn  io.ReadCloser
-	close   io.Closer
-	size    int64
-	name    string
-	closed  bool          // set if the file is closed
-	exit    chan struct{} // channel that will be closed when transfer is finished
-	withBuf bool          // is using a buffered in
+	mu       sync.Mutex // mutex protects these values
+	in       io.Reader
+	ctx      context.Context // current context for transfer - may change
+	ci       *fs.ConfigInfo
+	origIn   io.ReadCloser
+	close    io.Closer
+	size     int64
+	name     string
+	closed   bool          // set if the file is closed
+	exit     chan struct{} // channel that will be closed when transfer is finished
+	withBuf  bool          // is using a buffered in
+	checking bool          // set if attached transfer is checking
 
 	tokenBucket buckets // per file bandwidth limiter (may be nil)
 
@@ -295,14 +296,24 @@ func (acc *Account) ServerSideTransferEnd(n int64) {
 	acc.stats.Bytes(n)
 }
 
+// serverSideEnd accounts for non specific server side data
+func (acc *Account) serverSideEnd(n int64) {
+	// Account for bytes unless we are checking
+	if !acc.checking {
+		acc.stats.BytesNoNetwork(n)
+	}
+}
+
 // ServerSideCopyEnd accounts for a read of n bytes in a sever side copy
 func (acc *Account) ServerSideCopyEnd(n int64) {
 	acc.stats.AddServerSideCopy(n)
+	acc.serverSideEnd(n)
 }
 
 // ServerSideMoveEnd accounts for a read of n bytes in a sever side move
 func (acc *Account) ServerSideMoveEnd(n int64) {
 	acc.stats.AddServerSideMove(n)
+	acc.serverSideEnd(n)
 }
 
 // DryRun accounts for statistics without running the operation
@@ -528,9 +539,8 @@ func (acc *Account) String() string {
 	)
 }
 
-// rcStats produces remote control stats for this file
-func (acc *Account) rcStats() (out rc.Params) {
-	out = make(rc.Params)
+// rcStats adds remote control stats for this file
+func (acc *Account) rcStats(out rc.Params) {
 	a, b := acc.progress()
 	out["bytes"] = a
 	out["size"] = b
@@ -552,8 +562,6 @@ func (acc *Account) rcStats() (out rc.Params) {
 	}
 	out["percentage"] = percentageDone
 	out["group"] = acc.stats.group
-
-	return out
 }
 
 // OldStream returns the top io.Reader
